requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
api-display-purposes==0.0.3
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
clarifai==10.0.1
clarifai-grpc==10.0.4
click==8.1.7
configparser==3.8.1
contextlib2==21.6.0
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
Django==5.0.1
django-jazzmin==2.6.0
EasyProcess==1.1
emoji==2.10.0
environs==8.0.0
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==2.10
instapy==0.6.16
joblib==1.3.2
jsonschema==2.6.0
llama-index==0.9.34
magic-filter==1.0.12
markdown-it-py==3.0.0
marshmallow==3.20.2
mdurl==0.1.2
MeaningCloud-python==2.0.0
MouseInfo==0.1.3
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.8.0
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plyer==2.1.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pypdf==4.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.14
python-telegram-bot==20.7
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
PyVirtualDisplay==3.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
schema==0.7.5
selenium==4.16.0
semantic-version==2.10.0
setuptools-rust==1.8.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
tritonclient==2.41.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webdriverdownloader==1.1.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
gunicorn